+++ nutype_macros/src/integer/gen/mod.rs
@@ -63,17 +63,17 @@ where
         let validations: TokenStream = validators
             .iter()
             .map(|validator| match validator {
-                IntegerValidator::Max(max) => {
+                IntegerValidator::LessOrEqual(max) => {
                     quote!(
                         if val > #max {
-                            return Err(#error_name::MaxViolated);
+                            return Err(#error_name::LessOrEqualViolated);
                         }
                     )
                 }
-                IntegerValidator::Min(min) => {
+                IntegerValidator::GreaterOrEqual(min) => {
                     quote!(
                         if val < #min {
-                            return Err(#error_name::MinViolated);
+                            return Err(#error_name::GreaterOrEqualViolated);
                         }
                     )
                 }