+++ test_suite/tests/ui/integer/validate/min_vs_max.rs
@@ -1,6 +1,0 @@
-use nutype::nutype;
-
-#[nutype(validate(max = 0, min = 20))]
-pub struct Amount(i64);
-
-fn main() {}