+++ test_suite/tests/ui/integer/visibility/private_error.rs
@@ -1,7 +1,7 @@
 mod encapsulated {
     use nutype::nutype;
 
-    #[nutype(validate(min = 0, max = 100))]
+    #[nutype(validate(greater_or_equal = 0, less_or_equal = 100))]
     struct Percentage(i32);
 }